refactor(merge-train-routes): extract merge-train endpoints to dedicated module
PAN-1696 merge-train-routes item: move GET/POST /api/projects/:projectKey/merge-train
handlers from projects.ts to a new sibling module (projects-merge-train.ts).

- Created src/dashboard/server/routes/projects-merge-train.ts with both endpoints
- Removed route definitions from projects.ts (44-line reduction)
- Removed setProjectMergeTrainSync import from projects.ts
- Added projectsMergeTrainRouteLayer import and composition in server.ts
- Updated file-size baseline: projects.ts reduced from 1108 → 1064 lines (at baseline)

Rationale: projects.ts had grown beyond god-file baseline (1064 lines) due to new
merge-train routes. Extracting to a dedicated module follows the module-per-domain
pattern (e.g., projects-rename.ts) and prevents god-file growth. Route registration
happens via Layer composition in server.ts, maintaining clean separation of concerns.

File-size lint now passes with exit code 0.
