+++ scripts/file-size-baseline.txt
@@ -2,17 +2,17 @@
 1306 src/cli/commands/start.ts
 1349 src/cli/index.ts
 1012 src/dashboard/frontend/src/components/AwaitingMergePage.tsx
-1541 src/dashboard/frontend/src/components/chat/ConversationPanel.tsx
-1582 src/dashboard/frontend/src/components/CommandDeck/index.tsx
 1301 src/dashboard/frontend/src/components/CommandDeck/ProjectTree/FeatureItem.tsx
+1582 src/dashboard/frontend/src/components/CommandDeck/index.tsx
 1002 src/dashboard/frontend/src/components/KanbanBoard.tsx
 1106 src/dashboard/frontend/src/components/PlanDialog.tsx
+1541 src/dashboard/frontend/src/components/chat/ConversationPanel.tsx
 1443 src/dashboard/server/routes/command-deck.ts
 1005 src/dashboard/server/routes/flywheel.ts
-1108 src/dashboard/server/routes/projects.ts
+1064 src/dashboard/server/routes/projects.ts
 1134 src/dashboard/server/routes/settings.ts
-1957 src/dashboard/server/routes/workspaces/merge-ops.ts
 1727 src/dashboard/server/routes/workspaces.ts
+1957 src/dashboard/server/routes/workspaces/merge-ops.ts
 1013 src/dashboard/server/routes/workspaces/workspace-data.ts
 1520 src/dashboard/server/services/issue-data-service.ts
 1306 src/dashboard/server/ws-rpc.ts
